+++ game/components/base/light/s_noise.cpp
@@ -38,7 +38,7 @@ void System::Noise::init(Config &c, Scene *s, EntityID eid) {
     Component::Noise* noise = s->getComponent<Component::Noise>(eid);
     Component::Texture* tex = s->getComponent<Component::Texture>(eid);
     
-    noise->size = tex->size;
+    noise->size = tex->transform.size;
     noise->noise_data = std::vector<ui8>(noise->size.x * noise->size.y);
     
     Math::perlinNoise(noise->size, Vec2(0,0), noise->freq, noise->levels, noise->noise_data.data(), true);
@@ -54,8 +54,8 @@ void System::Noise::update(Config &c) {
         if (not noise->enabled)
             continue;
         
-        if (noise->size != tex->size) {
-            noise->size = tex->size;
+        if (noise->size != tex->transform.size) {
+            noise->size = tex->transform.size;
             noise->noise_data = std::vector<ui8>(noise->size.x * noise->size.y);
         }
         
@@ -114,7 +114,7 @@ void System::Noise::render(Config &c) {
             vertices[3*4+2] = vertices[1*4+2];
         }
         
-        Rect2 dst = Rect2 (tex->render_pos + noise->offset + ((tex->offset.size() > i) ? tex->offset[i] : Vec2(0, 0)), noise->size);
+        Rect2 dst = Rect2 (tex->transform.pos + noise->offset + ((tex->offset.size() > i) ? tex->offset[i] : Vec2(0, 0)), noise->size);
         
         noise->mask_tex.gl_id = tex->data.gl_id;
         noise->mask_tex.model = Graphics::Renderer::matModel2D(dst.pos - Vec2(BORDER_WIDTH, BORDER_WIDTH), dst.size);